04-datascience/requirements.txt
dockerspawner
jupyterhub-firstuseauthenticator
